splitting: make holdout private
